+++ baybe/targets/base.py
@@ -1,6 +1,10 @@
 """Base functionality for all BayBE targets."""
+
+from __future__ import annotations
+
 import warnings
 from abc import ABC, abstractmethod
+from typing import TYPE_CHECKING
 
 import pandas as pd
 from attrs import define, field
@@ -12,6 +16,9 @@ from baybe.utils import (
     unstructure_base,
 )
 
+if TYPE_CHECKING:
+    from baybe.objective import Objective
+
 
 @define(frozen=True)
 class Target(ABC, SerialMixin):
@@ -22,6 +29,12 @@ class Target(ABC, SerialMixin):
 
     name: str = field()
     """The name of the target."""
+
+    def to_objective(self) -> Objective:
+        """Create a single-task objective from the target."""
+        from baybe.objective import Objective
+
+        return Objective(targets=[self], mode="SINGLE")
 
     @abstractmethod
     def transform(self, data: pd.DataFrame) -> pd.DataFrame: